Spring Cleaning
Renamed the `member` route, splits the Directory controller.
