+++ assets/js/fos_js_routes.json
@@ -1,1 +1,1 @@
-{"base_url":"","routes":{"member":{"tokens":[["variable","\/","[^\/]++","localIdentifier"],["text","\/directory\/member"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"verify_address_data":{"tokens":[["text","\/directory\/verify-address-data"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"map_search":{"tokens":[["text","\/directory\/map-search"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"map_data":{"tokens":[["text","\/directory\/map-data"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"export_by_location":{"tokens":[["text","\/directory\/export\/by-location"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]}},"prefix":"","host":"localhost","port":"","scheme":"http","locale":[]}
+{"base_url":"","routes":{"member_show":{"tokens":[["variable","\/","[^\/]++","localIdentifier",true],["text","\/directory\/member"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"verify_address_data":{"tokens":[["text","\/directory\/verify-address-data"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"map_search":{"tokens":[["text","\/directory\/map-search"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"map_data":{"tokens":[["text","\/directory\/map-data"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]},"export_by_location":{"tokens":[["text","\/directory\/export\/by-location"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":[],"schemes":[]}},"prefix":"","host":"localhost","port":"","scheme":"http","locale":[]}
