+++ templates/directory/_row.html.twig
@@ -1,49 +1,49 @@
 <tr>
-  <td class="text-center" data-order="{{ record.localIdentifier }}">{{ record.localIdentifierShort }}</td>
+  <td class="text-center" data-order="{{ member.localIdentifier }}">{{ member.localIdentifierShort }}</td>
   <td width="32">
-    <a href="{{ path('member', {localIdentifier: record.localIdentifier}) }}"><img src="{{ record.photoUrl }}" class="img-fluid" style="max-width:32px;" /></a>
+    <a href="{{ path('member_show', {localIdentifier: member.localIdentifier}) }}"><img src="{{ member.photoUrl }}" class="img-fluid" style="max-width:32px;" /></a>
   </td>
-  <td data-order="{{ record.lastName }} {{ record.firstName }}">
-    <a href="{{ path('member', {localIdentifier: record.localIdentifier}) }}">{{ record.displayName }} </a>
+  <td data-order="{{ member.lastName }} {{ member.firstName }}">
+    <a href="{{ path('member_show', {localIdentifier: member.localIdentifier}) }}">{{ member.displayName }} </a>
     {% include 'member/_badges.html.twig' %}
   </td>
   {% if show_status is defined and show_status %}
   <td>
-    {{ record.status.label }}
+    {{ member.status.label }}
   </td>
   {% endif %}
   <td>
-    {{ record.classYear|default('-') }}
+    {{ member.classYear|default('-') }}
   </td>
   <td>
-    {% if record.primaryEmail %}
-    <a href="mailto:{{ record.primaryEmail }}">{{ record.primaryEmail }}</a>
+    {% if member.primaryEmail %}
+    <a href="mailto:{{ member.primaryEmail }}">{{ member.primaryEmail }}</a>
     {% endif %}
   </td>
   <td>
     <ul class="list-inline">
-      {% if record.linkedinUrl %}
-      <li class="list-inline-item m-0"><a href="{{ record.linkedinUrl }}" target="_blank"><i class="fab fa-linkedin fa-lg"></i></a></li>
+      {% if member.linkedinUrl %}
+      <li class="list-inline-item m-0"><a href="{{ member.linkedinUrl }}" target="_blank"><i class="fab fa-linkedin fa-lg"></i></a></li>
       {% else %}
       <li class="list-inline-item m-0"><i class="fab fa-linkedin fa-lg" style="opacity:0.2"></i></li>
       {% endif %}
-      {% if record.facebookIdentifier %}
-      <li class="list-inline-item m-0"><a href="https://facebook.com/{{ record.facebookIdentifier }}" target="_blank"><i class="fab fa-facebook fa-lg"></i></a></li>
+      {% if member.facebookIdentifier %}
+      <li class="list-inline-item m-0"><a href="https://facebook.com/{{ member.facebookIdentifier }}" target="_blank"><i class="fab fa-facebook fa-lg"></i></a></li>
       {% else %}
       <li class="list-inline-item m-0"><i class="fab fa-facebook fa-lg" style="opacity:0.2"></i></li>
       {% endif %}
     </ul>
   </td>
   <td class="text-nowrap">
-    {{ record.primaryTelephoneNumber }}
+    {{ member.primaryTelephoneNumber }}
   </td>
   <td>
-    {% if record.mailingAddressLine1 %}
-    {{ record.mailingAddressLine1 }}<br />
-    {% if record.mailingAddressLine2 %}{{ record.mailingAddressLine2 }}<br />{% endif %}
-    {{ record.mailingCity }}, {{ record.mailingState }} {{ record.mailingPostalCode }}<br />
-    {%- if record.mailingCountry != 'United States' and record.mailingCountry != 'US' %}
-    {{ record.mailingCountry }}
+    {% if member.mailingAddressLine1 %}
+    {{ member.mailingAddressLine1 }}<br />
+    {% if member.mailingAddressLine2 %}{{ member.mailingAddressLine2 }}<br />{% endif %}
+    {{ member.mailingCity }}, {{ member.mailingState }} {{ member.mailingPostalCode }}<br />
+    {%- if member.mailingCountry != 'United States' and member.mailingCountry != 'US' %}
+    {{ member.mailingCountry }}
     {% endif %}
     {% endif %}
   </td>
