+++ templates/directory/directory.html.twig
@@ -31,7 +31,7 @@
             </tr>
           </thead>
           <tbody>
-            {% for record in records %}
+            {% for member in members %}
             {% include 'directory/_row.html.twig' %}
             {% endfor %}
           </tbody>
