--- templates/directory/facebook.html.twig
@@ -1,20 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}{{ parent() }} - Facebook{% endblock %}
-
-{% block stylesheets -%}
-{{ parent() }}
-{% endblock %}
-
-{% block body %}
-<div class="h4">Facebook <sup><span class="badge badge-secondary">Beta</span></sup></div>
-{% for record in records %}
-  {% if record.facebookIdentifier %}
-  <a href="https://facebook.com/{{ record.facebookIdentifier }}"><img src="{{ record.photoUrl }}" data-toggle="tooltip" title="{{ record.displayName }}" class="img-fluid m-1" style="max-width:50px;" /></a>
-  {% endif %}
-{% endfor %}
-{% endblock %}
-
-{% block javascripts %}
-{{ parent() }}
-{% endblock %}
