+++ templates/member/_nav.html.twig
@@ -1,15 +1,18 @@
 {% set current_route = app.request.get('_route') %}
 <ul class="nav nav-tabs mb-3">
-  <li class="nav-item"><a class="nav-link {{ (current_route == 'member' or current_route == 'member_edit') ? 'active' }}" href="{{ path('member', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-id-badge"></i> Member Profile</a></li>
+  <li class="nav-item"><a class="nav-link {{ (current_route == 'member' or current_route == 'member_edit') ? 'active' }}" href="{{ path('member_show', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-id-badge"></i> Member Profile</a></li>
   {% if is_granted('ROLE_ADMIN') %}
-  <li class="nav-item"><a class="nav-link {{ current_route == 'member_change_log' ? 'active' }}" href="{{ path('member_change_log', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-history"></i> Change Log</a></li>
+  <li class="nav-item"><a class="nav-link {{ current_route == 'member_change_log' ? 'active' }}" href="{{ path('member_change_log', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-history"></i> Change Log</a></li>
   {% endif %}
-  {% if is_email_service_configured() and record.primaryEmail %}
-  <li class="nav-item"><a class="nav-link {{ current_route == 'member_email_subscription' ? 'active' }}" href="{{ path('member_email_subscription', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-envelope"></i> Email Subscriptions</a></li>
-  <li class="nav-item"><a class="nav-link {{ current_route == 'member_message' ? 'active' }}" href="{{ path('member_message', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-envelope-open-text"></i> Send Message</a></li>
+  {% if is_email_service_configured() and member.primaryEmail %}
+  <li class="nav-item"><a class="nav-link {{ current_route == 'member_email_subscription' ? 'active' }}" href="{{ path('member_email_subscription', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-envelope"></i> Email Subscriptions</a></li>
+  <li class="nav-item"><a class="nav-link {{ current_route == 'member_message' ? 'active' }}" href="{{ path('member_message', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-envelope-open-text"></i> Send Message</a></li>
+  {% elseif is_email_service_configured() and not member.primaryEmail %}
+  <li class="nav-item disabled"><a class="nav-link"><i class="fas fa-envelope"></i> Email Subscriptions</a></li>
+  <li class="nav-item disabled"><a class="nav-link"><i class="fas fa-envelope-open-text"></i> Send Message</a></li>
   {% endif %}
   {% if is_granted('ROLE_ADMIN') %}
-  <li class="nav-item"><a class="nav-link {{ current_route == 'member_donations' ? 'active' }}" href="{{ path('member_donations', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-hand-holding-usd"></i> Donations</a></li>
+  <li class="nav-item"><a class="nav-link {{ current_route == 'member_donations' ? 'active' }}" href="{{ path('member_donations', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-hand-holding-usd"></i> Donations</a></li>
   {% endif %}
-  <li class="nav-item"><a class="nav-link {{ current_route == 'member_vcard' ? 'active' }}" href="{{ path('member_vcard', {localIdentifier: record.localIdentifier}) }}"><i class="fas fa-address-card"></i> Download vCard</a></li>
+  <li class="nav-item"><a class="nav-link {{ current_route == 'member_vcard' ? 'active' }}" href="{{ path('member_vcard', {localIdentifier: member.localIdentifier}) }}"><i class="fas fa-address-card"></i> Download vCard</a></li>
 </ul>
