+++ templates/member/delete.html.twig
@@ -1,6 +1,6 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}{{ parent() }} - {{ record.displayName }} {% endblock %}
+{% block title %}{{ parent() }} - {{ member.displayName }} {% endblock %}
 
 {% block stylesheets -%}
 {{ parent() }}
@@ -9,7 +9,7 @@
 
 {% block body %}
 <div>
-  <div class="h4">{{ record.displayName }}</div>
+  <div class="h4">{{ member.displayName }}</div>
 
   {% include 'member/_nav.html.twig' %}
