+++ templates/search/index.html.twig
@@ -9,14 +9,14 @@
 {% set member = row[0] %}
 <div class="card mb-3 shadow" style="max-width:40em;">
   <div class="card-body">
-    <a href="{{ path('member', {localIdentifier: member.localIdentifier}) }}">{{ member }}</a> - {{ member.status.label }}
-    {% include 'member/_badges.html.twig' with {record: member} %}
+    <a href="{{ path('member_show', {localIdentifier: member.localIdentifier}) }}">{{ member }}</a> - {{ member.status.label }}
+    {% include 'member/_badges.html.twig' %}
   </div>
 </div>
 {% else %}
 <div class="card mb-3 shadow">
   <div class="card-body">
-    <p>No members matched your search criteria.</p>
+    <div class="alert alert-info m-0">No members matched your search criteria.</div>
   </div>
 </div>
 {% endfor %}
