+++ templates/tag/show.html.twig
@@ -17,7 +17,7 @@
   <div class="card-body">
     <ul>
       {% for member in tag.members %}
-      <li><a href="{{ path('member', {localIdentifier: member.localIdentifier}) }}">{{ member }}</a></li>
+      <li><a href="{{ path('member_show', {localIdentifier: member.localIdentifier}) }}">{{ member }}</a></li>
       {% endfor %}
     </ul>
   </div>
